backend/courses/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic[email]
python-dotenv
pandas
openpyxl
python-multipart
